requirements.txt
argcomplete==3.5.0
asgiref==3.8.1
black==24.8.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
commitizen==3.29.1
decli==0.6.2
distlib==0.3.8
Django==5.1.1
filelock==3.16.1
flake8==7.1.1
identify==2.6.1
Jinja2==3.1.4
MarkupSafe==2.1.5
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
pre-commit==3.8.0
prompt-toolkit==3.0.36
pycodestyle==2.12.1
pyflakes==3.2.0
PyYAML==6.0.2
questionary==2.0.1
sqlparse==0.5.1
termcolor==2.4.0
tomlkit==0.13.2
tzdata==2024.2
virtualenv==20.26.6
wcwidth==0.2.13

djangorestframework~=3.15.2
pandas~=2.2.3
